librats: support get collateral from evidence and url in sev-snp verifier
