+++ verifiers/sev-snp/verify_evidence.c
@@ -29,7 +29,11 @@ rats_verifier_err_t sev_snp_verify_evidence(rats_verifier_ctx_t *ctx,
 	rats_verifier_err_t err = RATS_VERIFIER_ERR_UNKNOWN;
 	snp_attestation_report_t *report = (snp_attestation_report_t *)(evidence->snp.report);
 	if (evidence->snp.vcek[0] == '\0') {
-		return RATS_VERIFIER_ERR_INVALID;
+		memset(evidence->snp.vcek, 0, VECK_MAX_SIZE);
+		err = sev_snp_get_vcek_der(report->chip_id, sizeof(report->chip_id),
+					   &report->platform_version, &evidence->snp);
+		if (err != RATS_ATTESTER_ERR_NONE)
+			return err;
 	}
 
 	/* Verify the hash value */
@@ -60,7 +64,7 @@ rats_verifier_err_t sev_snp_verify_evidence(rats_verifier_ctx_t *ctx,
 	if (!x509_ark) {
 		goto err;
 	}
-	char *vcek_ptr = evidence->snp.vcek;
+	uint8_t *vcek_ptr = evidence->snp.vcek;
 	x509_vcek = d2i_X509(NULL, (const unsigned char **)&(vcek_ptr), evidence->snp.vcek_len);
 	if (!x509_vcek) {
 		goto err;
